optimized for github

+++ web_manipulation.py
@@ -1,5 +1,3 @@
-from selenium import webdriver
-from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
@@ -7,6 +5,7 @@ from selenium.common.exceptions import TimeoutException, NoSuchElementException
 from selenium.webdriver.common.action_chains import ActionChains
 import time
 import pickle
+
 
 class SeleniumDriver:
     def __init__(self, driver):
@@ -32,6 +31,23 @@ class WebManipulation(SeleniumDriver):
         except TimeoutException:
             pass
 
+    def check_element_exists(self, xpath: str):
+        try:
+            # Wait for up to 10 seconds before throwing a TimeoutException
+            element = WebDriverWait(self.driver, 10).until(
+                EC.presence_of_element_located((By.XPATH, xpath))
+            )
+            # If we made it this far, the element must exist
+            return True
+        except TimeoutException:
+            print(f"TimeoutException: The element with the xpath '{xpath}' was not found within 10 seconds.")
+            return False
+        except NoSuchElementException:
+            print(f"NoSuchElementException: The element with the xpath '{xpath}' does not exist on the page.")
+            return False
+        except Exception as e:
+            print(f"Unexpected error occurred when trying to check the existence of the element with the xpath '{xpath}': {str(e)}")
+            return False
 
     def click_element(self, xpath: str):
         try:
@@ -39,7 +55,6 @@ class WebManipulation(SeleniumDriver):
             element = WebDriverWait(self.driver, 10).until(
                 EC.presence_of_element_located((By.XPATH, xpath))
             )
-            #element.click()
             self.driver.execute_script("arguments[0].click();", element)
         except TimeoutException:
             print(f"TimeoutException: The element with the xpath '{xpath}' was not found within 10 seconds.")
@@ -48,31 +63,43 @@ class WebManipulation(SeleniumDriver):
         except Exception as e:
             print(f"Unexpected error occurred when trying to click the element with the xpath '{xpath}': {str(e)}")
 
+    def write_into_element_with_actionchains(self, xpath: str, text: str, max_attempts: int = 3):
+        for attempt in range(max_attempts):
+            try:
+                # Wait for up to 10 seconds before throwing a TimeoutException
+                element = WebDriverWait(self.driver, 10).until(
+                    EC.presence_of_element_located((By.XPATH, xpath))
+                )
+                actions = ActionChains(self.driver)
+                actions.move_to_element(element).click().perform()
+                actions.reset_actions()
 
+                for char in text:
+                    actions.send_keys(char)
+                    time.sleep(1.5)  # adjust the sleep time as needed
+                actions.perform()
+                # Dispatch 'input' event
+                self.driver.execute_script("arguments[0].dispatchEvent(new Event('input'));", element)
+                # Dispatch 'change' event
+                self.driver.execute_script("arguments[0].dispatchEvent(new Event('change'));", element)
 
-    def write_into_element_with_actionchains(self, xpath: str, text: str):
-        try:
-            # Wait for up to 10 seconds before throwing a TimeoutException
-            element = WebDriverWait(self.driver, 10).until(
-                EC.presence_of_element_located((By.XPATH, xpath))
-            )
-            actions = ActionChains(self.driver)
-            actions.move_to_element(element).click()
-            for char in text:
-                actions.send_keys(char)
-                time.sleep(1.5)  # adjust the sleep time as needed
-            actions.perform()
-            # Dispatch 'input' event
-            self.driver.execute_script("arguments[0].dispatchEvent(new Event('input'));", element)
-            # Dispatch 'change' event
-            self.driver.execute_script("arguments[0].dispatchEvent(new Event('change'));", element)
-        except TimeoutException:
-            print(f"TimeoutException: The element with the xpath '{xpath}' was not found within 10 seconds.")
-        except NoSuchElementException:
-            print(f"NoSuchElementException: The element with the xpath '{xpath}' does not exist on the page.")
-        except Exception as e:
-            print(f"Unexpected error occurred when trying to write into the element with the xpath '{xpath}': {str(e)}")
+                # Validate if the written text is correct
+                entered_text = element.get_attribute("value")
+                if entered_text == text:
+                    break  # Text is correctly entered
+                else:
+                    print(f"The written text '{entered_text}' does not match the desired text '{text}'. Retrying...")
 
+            except TimeoutException:
+                print(f"TimeoutException: The element with the xpath '{xpath}' was not found within 10 seconds.")
+            except NoSuchElementException:
+                print(f"NoSuchElementException: The element with the xpath '{xpath}' does not exist on the page.")
+            except Exception as e:
+                print(f"Unexpected error occurred when trying to write into the element with the xpath '{xpath}': {str(e)}")
+
+            # If we've reached the maximum number of attempts, raise an exception
+            if attempt == max_attempts - 1:
+                raise Exception(f"Failed to correctly enter the text '{text}' into the element with xpath '{xpath}' after {max_attempts} attempts.")
 
     def write_into_element(self, xpath: str, text: str):
         try:
@@ -93,7 +120,7 @@ class WebManipulation(SeleniumDriver):
         except Exception as e:
             print(f"Unexpected error occurred when trying to write into the element with the xpath '{xpath}': {str(e)}")
 
-    def send_keys_to_element(self,xpath: str, keys: str):
+    def send_keys_to_element(self, xpath: str, keys: str):
         try:
             # Wait for up to 10 seconds before throwing a TimeoutException
             element = WebDriverWait(self.driver, 10).until(
@@ -149,7 +176,7 @@ class WebManipulation(SeleniumDriver):
         except Exception as e:
             print(f"Unexpected error occurred when trying to scroll to the bottom of the page: {str(e)}")
 
-    def get_element_text(self, xpath: str) ->str:
+    def get_element_text(self, xpath: str) -> str:
         try:
             # Wait for up to 10 seconds before throwing a TimeoutException
             element = WebDriverWait(self.driver, 10).until(
@@ -198,12 +225,11 @@ class WebManipulation(SeleniumDriver):
             print(f"Unexpected error occurred when trying to get '{attribute}' attribute of the element with the xpath '{xpath}': {str(e)}")
             return None
 
+
 class GetNewListings(SeleniumDriver):
-    is_first_run = False
 
     def __init__(self, driver):
         super().__init__(driver)
-
 
     def get_new_listings(self, path: str, page_last_number_btn: str, next_page_btn: str, listing_link: str) -> set:
         """
@@ -220,41 +246,32 @@ class GetNewListings(SeleniumDriver):
         wm = WebManipulation(self.driver)
         data_set = set()
         # get number of pages with listings
-        pagecount = int(wm.get_element_text(page_last_number_btn))
-        for page_index in range(pagecount):
-            if page_index != 0:
+        btn_exist = wm.check_element_exists(page_last_number_btn)
+        if btn_exist:
+            pagecount = int(wm.get_element_text(page_last_number_btn))
+            for page_index in range(pagecount):
+                if page_index != 0:
+                    time.sleep(1)
+                    wm.click_element(next_page_btn)
                 time.sleep(1)
-                #wm.scroll_to_element(f'//button[contains(@aria-label, "Idź do strony") and contains(text(), "{(page_index+1)}")]')
-                wm.click_element(next_page_btn)
-            time.sleep(1)
-            # scroll to bottom in order to load all listings
-            wm.scroll_to_bottom()
-            listings_count = wm.get_elements_count(listing_link)
-            for i in range(listings_count):
-                url_to_listing = wm.get_element_attribute(f'({listing_link})[{i+1}]', 'href')
-                data_set.add(url_to_listing)
+                # scroll to bottom in order to load all listings
+                wm.scroll_to_bottom()
+                listings_count = wm.get_elements_count(listing_link)
+                for i in range(listings_count):
+                    url_to_listing = wm.get_element_attribute(f'({listing_link})[{i+1}]', 'href')
+                    data_set.add(url_to_listing)
 
-        try:
-            with open(path, 'rb') as f:
-                compare_data = pickle.load(f)
-                new_data = data_set.difference(compare_data)
-                #print(f'not first run {path}')
-                #print(new_data)
-            with open(path, 'wb') as f:
-                pickle.dump(data_set, f)
-                return new_data
+            try:
+                with open(path, 'rb') as f:
+                    compare_data = pickle.load(f)
+                    new_data = data_set.difference(compare_data)
+                with open(path, 'wb') as f:
+                    pickle.dump(data_set, f)
+                    return new_data
 
-        except:
-            with open(path, 'wb') as f:
-                pickle.dump(data_set, f)
-                #print(f'first run {path}')
-                #print(data_set)
-                return data_set
-
-
-
-
-
-
-
-
+            except:
+                with open(path, 'wb') as f:
+                    pickle.dump(data_set, f)
+                    return data_set
+        else:
+            return data_set
